servicion creando cliente falta manejo de autenticacion

+++ gymstatsapirest/src/main/java/com/gymstatsapirest/repository/GeneroRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gymstatsapirest.model.Genero;
 
-public interface GeneroRepository extends JpaRepository<Genero, Integer>
+public interface GeneroRepository extends JpaRepository<Genero, Short>
 {
 
 }
